API commit 11/6
I was able to successfully create and test the API connection for both
my resources (user and animals) I can now successfully create, read,
update and delete both via the CLI.

To do this, I had to work on my ID and ID connections including the
user/creator, owner and the token - that appeared to be the source of
the bulk of my issues.  I troubleshooted with a few peers to confirm.
Additionally, I adjusted my Github connection to the repo (one for the
API, one for the client side) and I connected the server.js router to my
Animal schema which had given me a few issues.  As a next step, I need
to deploy on Heroku and confirm success there before I can begin working
on my frontend.

+++ app/routes/animal_routes.js
@@ -1,53 +1,74 @@
 // require express
-const express = require('express')
-// require crypto
-const crypto = require('crypto')
-// require error types
-const errors = require('../../lib/custom_errors')
+// require error types and their references
 // require 404 from the errors
-const handle404 = require('../../lib/custom_errors')
+const express = require('express')
+const customErrors = require('../../lib/custom_errors')
+const handle404 = customErrors.handle404
 
 // create an express router
-const router = express.Router()
-
-// reference the parameters variable and files
-const BadParamsError = errors.BadParamsError
-const BadCredentialsError = errors.BadCredentialsError
+// require passport
 // require the animal schema from models
+const router = express.Router()
+const passport = require('passport')
 const Animal = require('../models/animal')
 
+// define the requireToken variable via passport
+// define the requireOwnership variable via the custom_errors folder
+const requireToken = passport.authenticate('bearer', { session: false })
+const requireOwnership = customErrors.requireOwnership
+
 // CREATE a new animal
-router.post('/animals', (req, res, next) => {
-  // set owner of new example to be current user
+router.post('/animals', requireToken, (req, res, errors) => {
   req.body.animal.owner = req.user.id
-  Animal.create(req.body.example)
-    // respond to succesful `create` with status 201 and JSON of new "example"
+  Animal.create(req.body.animal)
     .then(animal => {
-      res.status(201).json({ example: animal.toObject() })
+      res.status(201).json({ animal: animal.toObject() })
     })
-    .catch(next)
+    .catch(errors)
 })
 
 // UPDATE an existing animal
-router.patch('/animals/:id', (req, res, next) => {
+router.patch('/animals/:id', requireToken, (req, res, errors) => {
+  delete req.body.animal.owner
   Animal.findById(req.params.id)
     .then(handle404)
     .then(animal => {
+      requireOwnership(req, animal)
       return animal.updateOne(req.body.animal)
     })
     .then(() => res.sendStatus(204))
-    .catch(next)
+    .catch(errors)
 })
 
 // DESTROY an existing animal (ah!)
-router.delete('/animals/:id', (req, res, next) => {
+router.delete('/animals/:id', requireToken, (req, res, errors) => {
   Animal.findById(req.params.id)
     .then(handle404)
     .then(animal => {
+      requireOwnership(req, animal)
       animal.deleteOne()
     })
     .then(() => res.sendStatus(204))
-    .catch(next)
+    .catch(errors)
+})
+
+// INDEX existing animals
+router.get('/animals', requireToken, (req, res, errors) => {
+  Animal.find()
+    .then(animal => {
+      return animal.map(animal => animal.toObject())
+    })
+    .then(animal => res.status(200).json({ animal: animal }))
+    .catch(errors)
+})
+
+// SHOW existing animals
+router.get('/animals/:id', requireToken, (req, res, errors) => {
+  req.body.animal.owner = req.user.id
+  Animal.findById(req.params.id)
+    .then(handle404)
+    .then(animal => res.status(200).json({ animal: animal.toObject() }))
+    .catch(errors)
 })
 
 module.exports = router
